@@ -1,7 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
-// import App from "./App";
-import App from "./App_functional";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+import App from "./App";
+//import App from "./backup/App_classbased";
+
+import { createRoot } from "react-dom/client";
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(<App />);
